# src/swagger/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"IReadingChallenge": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"title": {
						"type": "string",
						"nullable": true
					},
					"type": {
						"type": "string",
						"enum": [
							"WEEKLY",
							"MONTHLY",
							"ANNUAL"
						]
					},
					"startDate": {
						"type": "string",
						"format": "date-time"
					},
					"endDate": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"progress": {
						"type": "number",
						"format": "double"
					},
					"goal": {
						"type": "number",
						"format": "double"
					},
					"timeframe": {
						"type": "string"
					},
					"hasEnded": {
						"type": "boolean"
					},
					"userId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id",
					"title",
					"type",
					"startDate",
					"endDate",
					"progress",
					"goal",
					"timeframe",
					"hasEnded",
					"userId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IAuthor": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pick_IBook.id-or-title-or-coverPicture-or-authors-or-rating_": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"title": {
						"type": "string"
					},
					"coverPicture": {
						"type": "string"
					},
					"authors": {
						"items": {
							"$ref": "#/components/schemas/IAuthor"
						},
						"type": "array"
					},
					"rating": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id",
					"title",
					"coverPicture",
					"authors",
					"rating"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"IReadingChallengeBook": {
				"$ref": "#/components/schemas/Pick_IBook.id-or-title-or-coverPicture-or-authors-or-rating_"
			},
			"IReadingChallengeWithBooks": {
				"allOf": [
					{
						"$ref": "#/components/schemas/IReadingChallenge"
					},
					{
						"properties": {
							"books": {
								"items": {
									"$ref": "#/components/schemas/IReadingChallengeBook"
								},
								"type": "array"
							}
						},
						"required": [
							"books"
						],
						"type": "object"
					}
				]
			},
...
			"Pick_IBook.Exclude_keyofIBook.authors-or-genres__": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"country": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"coverPicture": {
						"type": "string"
					},
					"rating": {
						"type": "number",
						"format": "double"
					},
					"isbn": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"publishDate": {
						"type": "string",
						"format": "date-time"
					},
					"format": {
						"type": "string",
						"enum": [
							"PAPERBACK",
							"HARDCOVER",
							"EBOOK"
						]
					},
					"language": {
						"type": "string"
					},
					"numOfPages": {
						"type": "number",
						"format": "double"
					},
					"pdfLink": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"country",
					"title",
					"coverPicture",
					"rating",
					"isbn",
					"description",
					"publishDate",
					"format",
					"language",
					"numOfPages",
					"pdfLink"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_IBook.authors-or-genres_": {
				"$ref": "#/components/schemas/Pick_IBook.Exclude_keyofIBook.authors-or-genres__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"IBookWithoutAuthorsAndGenres": {
				"$ref": "#/components/schemas/Omit_IBook.authors-or-genres_"
			},
...
			"Duration": {
				"enum": [
					"WEEKLY",
					"MONTHLY",
					"ANNUAL"
				],
				"type": "string"
			},
			"CreateReadingChallengeDto": {
				"properties": {
					"title": {
						"type": "string"
					},
					"type": {
						"$ref": "#/components/schemas/Duration"
					},
					"userId": {
						"type": "number",
						"format": "double"
					},
					"goal": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"title",
					"type",
					"userId",
					"goal"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateReadingChallengeDto": {
				"properties": {
					"title": {
						"type": "string"
					},
					"goal": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"title",
					"goal"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "grad-proj-core",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/reading-challenges": {
			"get": {
				"operationId": "GetAllReadingChallenges",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"items": {
												"$ref": "#/components/schemas/IReadingChallengeWithBooks"
											},
											"type": "array"
										},
										{}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Reading Challenges"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateReadingChallenge",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/IReadingChallenge"
										},
										{}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Reading Challenges"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateReadingChallengeDto"
							}
						}
					}
				}
			}
		},
		"/reading-challenges/{readingChallengeId}": {
			"get": {
				"operationId": "GetReadingChallengeById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/IReadingChallengeWithBooks"
										},
										{}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Reading Challenges"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "readingChallengeId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdateReadingChallengeDetails",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/IReadingChallenge"
										},
										{}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Reading Challenges"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "readingChallengeId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateReadingChallengeDto"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteReadingChallenge",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/IReadingChallenge"
										},
										{}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Reading Challenges"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "readingChallengeId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/reading-challenges/{readingChallengeId}/books": {
			"get": {
				"operationId": "GetBooksByReadingChallengeId",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"items": {
												"$ref": "#/components/schemas/IBookWithoutAuthorsAndGenres"
											},
											"type": "array"
										},
										{}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Reading Challenges"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "readingChallengeId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/reading-challenges/{userId}/add-book/{bookId}": {
			"patch": {
				"operationId": "AddBookToUserReadingChallenges",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"items": {
												"$ref": "#/components/schemas/IReadingChallengeWithBooks"
											},
											"type": "array"
										},
										{}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Reading Challenges"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "path",
						"name": "bookId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/reading-challenges/{userId}/remove-book/{bookId}": {
			"patch": {
				"operationId": "DeleteBookFromUserReadingChallenges",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"items": {
												"$ref": "#/components/schemas/IReadingChallengeWithBooks"
											},
											"type": "array"
										},
										{}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Reading Challenges"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "path",
						"name": "bookId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}